Continue helper function consolidation
