+++ iati/tests/news_functional_tests.py
@@ -4,11 +4,9 @@ TODO:
     Refactor most of these tests out into base functional tests.
 
 """
-import os
 import pytest
-from django.utils.text import slugify
-from base_functional_tests import find_and_click_add_button, find_and_click_toggle_button, fill_content_editor_block, click_obscured, view_live_page
-from iati.urls import ADMIN_SLUG
+from tests import helper_functions
+from base_functional_tests import find_and_click_add_button, find_and_click_toggle_button, fill_content_editor_block
 
 
 NEWS_INDEX_PAGE = {
@@ -27,18 +25,6 @@ TEST_CATEGORY = "Test IATI news category"
 H2 = {'content': 'H2 heading', 'button': 'H2', 'id': 'content_editor_en-{}-value'}
 H3 = {'content': 'H3 heading', 'button': 'H3', 'id': 'content_editor_en-{}-value'}
 H4 = {'content': 'H4 heading', 'button': 'H4', 'id': 'content_editor_en-{}-value'}
-
-
-def navigate_to_default_page_cms_section(admin_browser, default_page_title):
-    """Navigate to the news page section of the CMS.
-
-    Args:
-        default_page_title (str): The title of a default page modelself.
-
-    """
-    admin_browser.click_link_by_text('Pages')
-    admin_browser.find_by_xpath('//span[@class="icon icon-arrow-right "]').click()
-    admin_browser.find_by_text(default_page_title).click()
 
 
 def enter_page_content(admin_browser, tab_name, cms_field, cms_content):
@@ -64,75 +50,19 @@ def publish_page(admin_browser):
     admin_browser.find_by_text('Publish').click()
 
 
-def create_news_child_page(admin_browser, page_type, page_title):
-    """Create a child page in the CMS.
-
-    Args:
-        page_type (str): The verbose name of the page model type you want to click on.
-        page_title (str): The title of the page you are editing.
-
-    """
-    navigate_to_default_page_cms_section(admin_browser, 'News')
-    admin_browser.find_by_text('Add child page').click()
-    admin_browser.find_by_text(page_type).click()
-    enter_page_content(admin_browser, 'English', 'title_en', page_title)
-    enter_page_content(admin_browser, 'Promote', 'slug_en', slugify(page_title))
-    publish_page(admin_browser)
-
-
-def edit_page_header(admin_browser, page_title, cms_field, cms_content):
-    """Edit a page by adding content via the CMS.
-
-    Args:
-        page_title (str): The title of the page you want to edit.
-        cms_field (str): The name of the CMS field you want to enter content into.
-        cms_content (str): The text content you want to add to the CMS field.
-
-    TODO:
-        Rename this function to avoid confusion with content editor tests.
-
-    """
-    admin_browser.find_by_text(page_title).click()
-    enter_page_content(admin_browser, 'English', cms_field, cms_content)
-    publish_page(admin_browser)
-    view_live_page(admin_browser, page_title)
-
-
-def scroll_to_bottom_of_page(admin_browser):
-    """Scroll to the bottom of a page."""
-    admin_browser.driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")
-
-
-def reveal_content_editor(admin_browser, button, element_count):
-    """Open content editor if it is not already in view.
-
-    Args:
-        button (str): The displayed text name of the button you want to click.
-        element_count (str): The element counter value of the content editor Streamfield block.
-
-    TODO:
-        Decide whether on convert element_count to int on assignment of variable.
-
-    """
-    if not admin_browser.find_by_text(button).visible and int(element_count) > 0:
-        toggle = admin_browser.find_by_xpath('//div[@id="content_editor_en-{}-appendmenu"]/a'.format(int(element_count) - 1))[0]
-        admin_browser.driver.execute_script("arguments[0].click();", toggle.__dict__['_element'])
-        scroll_to_bottom_of_page(admin_browser)
-
-
 @pytest.mark.django_db
 class TestNewsPage():
     """A container for tests to check functionality of news pages and child pages."""
 
     def test_can_edit_news_page_heading(self, admin_browser):
         """Check that an existing news page heading can be edited."""
-        navigate_to_default_page_cms_section(admin_browser, 'News')
-        edit_page_header(admin_browser, NEWS_INDEX_PAGE['title'], 'heading_en', NEWS_INDEX_PAGE['heading'])
+        helper_functions.navigate_to_default_page_cms_section(admin_browser, 'News')
+        helper_functions.edit_page_header(admin_browser, NEWS_INDEX_PAGE['title'], 'heading_en', NEWS_INDEX_PAGE['heading'])
         assert admin_browser.find_by_text(NEWS_INDEX_PAGE['heading'])
 
     def test_can_edit_news_page_excerpt(self, admin_browser):
         """Check that an existing news page excerpt can be edited."""
-        edit_page_header(admin_browser, NEWS_INDEX_PAGE['title'], 'excerpt_en', NEWS_INDEX_PAGE['excerpt'])
+        helper_functions.edit_page_header(admin_browser, NEWS_INDEX_PAGE['title'], 'excerpt_en', NEWS_INDEX_PAGE['excerpt'])
         assert admin_browser.find_by_text(NEWS_INDEX_PAGE['excerpt'])
 
 
@@ -142,14 +72,14 @@ class TestNewsIndexChildPages():
 
     def test_can_create_news_child_pages(self, admin_browser):
         """Check that when an news index child page is created it appears in the website."""
-        create_news_child_page(admin_browser, NEWS_PAGE['page_type'], NEWS_PAGE['title'])
-        view_live_page(admin_browser, NEWS_PAGE['title'])
+        helper_functions.create_news_child_page(admin_browser, NEWS_PAGE['page_type'], NEWS_PAGE['title'])
+        helper_functions.view_live_page(admin_browser, NEWS_PAGE['title'])
         assert not admin_browser.is_text_present('Home')
         assert admin_browser.title == NEWS_PAGE['title']
 
     def test_can_edit_news_child_page_heading(self, admin_browser):
         """Check that news index child page headings can be edited."""
-        edit_page_header(admin_browser, NEWS_PAGE['title'], 'heading_en', NEWS_PAGE['heading'])
+        helper_functions.edit_page_header(admin_browser, NEWS_PAGE['title'], 'heading_en', NEWS_PAGE['heading'])
         assert admin_browser.find_by_text(NEWS_PAGE['heading'])
 
     @pytest.mark.parametrize('header', [
@@ -165,7 +95,7 @@ class TestNewsIndexChildPages():
         find_and_click_add_button(admin_browser, header['button'].lower())
         fill_content_editor_block(admin_browser, header['button'].lower(), " input", header['content'])
         publish_page(admin_browser)
-        view_live_page(admin_browser, NEWS_PAGE['title'])
+        helper_functions.view_live_page(admin_browser, NEWS_PAGE['title'])
         assert admin_browser.is_text_present(header['content'])
     # TODO: Find out why this is consistently failing
     # def test_news_category_filter(self, admin_browser):
